delete unit test for objects router; final fixes for variations router unit tests
